Update AIModule.cpp
The AI no longer considers the current position of a unit as save in terms of picking up gear or pre-priming a grenade, when it sees an enemy unit from that position or thinks it was seen by the enemy this turn.

Scoring-mechanism for peeking-position now gets a bonus for overlapping with the assumed path of enemies as to make it more likely to look at where they're actually likely to come from.