Update AIModule.cpp
With enabled dynamic aggressiveness, the AI will now attempt desperation-attacks if their numbers are very low to spare you looking for them in some closet.

With enabled dynamic aggressiveness, the AI will now always raise aggressiveness to at least 2 when they think one of their units is cornered, regardless of numerical advantages.

Aggressiveness 3 no longer automatically works like leeroy-mode. Instead units will spend the majority of their time-units to move towards the enemy and then look for opportune cover. Units who think they have been seen will also look for cover sooner.

AI units will no longer turn on the spot if doing so would exceed the TUs they intended to save for hiding.

Energy is now also considered when checking if a unit has enough time-units to still attack from a certain tile.

Direct peeking-behavior is now generally much bolder on all aggressiveness-levels. If a unit has enough TUs and energy to attack from a position from where they think they could see and attack enemy from, they will risk going there, even if it exceeds the TU-budget they wanted to save for hiding.

Direct peeking also no longer just considers the assumed closest unit by walking distance but instead all units who's locations the AI thinks it knows.

On Aggressiveness 2 and 3, even for indirect peeking that is considered unlikely to reveal an enemy, the AI will now expend way more TUs. Either as much as it barely will have enough for the cheapest attack or if the cheapest attack is too expensive, up to 70% of their TUs and energy.

Seeking for the best cover now works on all aggressiveness-levels. The main difference is how many TUs are spent for other tasks beforehand.

The AI will no longer skip peeking if it wants to prime a grenade but doesn't consider itself in an already safe location.

On aggressiveness 1 the AI now also tries to use the best compromise between uncovering many tiles and saving TUs for indirect peeking. Direct peeking, as mentioned is much less restricted.

Fixed that the AI thought opening doors costs only 4 TUs when it actually costs 8.

When all cover-options seem equally good or bad the AI now uses distance to enemy as a tie-breaker. If it thinks it was discovered or knows it has given away it's position, it'll try to increase distance. If it thinks it wasn't discovered yet, it'll try to decrease distance.

For direct-peeking the AI now determines the direction to look at before moving.

For deciding if a tile is worth to indirectly peek from, the AI only considers the best direction to look at instead of the sum of all directions. This prevents indecisive peeking-behavior like alternating between two tiles.

On aggressiveness 3 new-tile discovery is irrelevant for the decision to go somewhere, as the idea is to move all units towards a location as quickly as possible.

Aggressiveness 2 can now be considered to be mostly about map-control. If no units are found, the AI will head towards positions from where it can see the most tiles.