General AI improvements
Immobile units are always allowed to check if they can attack something.

Rewrote the way of how the AI determines where to move to perform an attack from.

AI will no longer attempt to attack with completely inefficient weapons and will also not move to a tile from which it could only attack with completely inefficient weapons.

Fixed an issue where units that already could attack at the beginning of their turn wouldn't do so.

A unit that has a psi- or blaster-attack available will still wait until after moving to consider attacks with other weapons than the psi or blaster.

Units that already see a unit will no longer turn around in order to spot units.

Cyberdisks and similar units no longer behave as if they were melee-units. The idea had some charm but overall it caused more harm than good.

The AI will now avoid walking towards enemies who have the high-ground of an elevator because the way vision works in this game means that the units ontop of the elevator will be guaranteed to see the units below first so that no mutual surprise rule will apply. This does not include units with the Leeroy-flag.

The AI will now consider the time-units it needs to turn when determining whether it has enough TUs to perform an attack from it's current tile. This way it won't waste TUs it would otherwise be able to use for hiding.

The AI should now be better at avoiding reaction-fire.

Cut down the amount of tiles that flying units consider for indirect peeking. They will now only consider tiles  with a floor, tiles on the path towards their supposed enemy-location and tiles directly above them for doing so. This reduces turn-times when there are many flying units.

Indirect peek-directions only double their score for containing the path towards an enemy instead of multiplying the score by the visible path-length. This shall help the AI to scout their flanks more frequently.

AI will no longer save the TUs for stepping back one tile if the tile they'd step back to is visible to their enemies anyways.