Fix RandomWound issue
Fix for incompatibility of previous version with Mods that use "RandomWound":

Reverted the "RandomWound"-stat back to boolean and made a new stat "RandomWoundType".
The default for "RandomWoundType" is the old behavior of "RandomWound".
In order to use different RandomWoundTypes it has to be used in conjunction with RandomWound = true.