Use ammo-damage-type
Now properly using the damage-type of the ammo for damage predictions instead of the damage-type of the weapon itself.
Damage-type is not mandatory for weapons if they use ammo, so in some mods it wasn't set for all weapons. This lead to assuming they had no damage-type and thus no damage despite the ammo holding that kind of information.
=> This fixed an issue seen in X-Piratez where units on autoplay would not use their weapon.